Reorganize internals & readme
