+++ lib/thrifter/extensions/ping.rb
