+++ lib/thrifter/extensions/retriable.rb
